Moving Twitter helpers to separate file.
