+++ resources/assets/components/utilities/Share/Share.js
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import { handleTwitterShareClick } from '../../../helpers';
+import { handleTwitterShareClick } from '../../../helpers/twitter';
 import { handleFacebookShareClick } from '../../../helpers/facebook';
 
 import './share.scss';
