+++ resources/assets/helpers/index.js
@@ -608,21 +608,6 @@ export function openDialog(href, callback, width = 550, height = 420) {
 }
 
 /**
- * Share a link by opening a Twitter share prompt.
- *
- * @param  {String} href
- * @param  {String} quote
- */
-export function showTwitterSharePrompt(href, quote = '', callback) {
-  const intent = makeUrl('https://twitter.com/intent/tweet', {
-    url: href,
-    text: quote,
-  });
-
-  openDialog(intent, callback);
-}
-
-/**
  * Handle click event from a Snapchat share button
  *
  * @param {String} href
@@ -636,24 +621,6 @@ export function handleSnapchatShareClick(href, trackingData) {
     label: 'snapchat',
     context: { ...trackingData, url: href },
   });
-}
-
-/**
- * Handle click event from a Twitter share button
- *
- * @param {String} href
- * @param {Object} trackingData
- * @param {String} quote
- */
-export function handleTwitterShareClick(href, trackingData, quote = '') {
-  trackAnalyticsEvent('clicked_share_twitter', {
-    action: 'button_clicked',
-    category: EVENT_CATEGORIES.socialShare,
-    label: 'twitter',
-    context: { ...trackingData, url: href },
-  });
-
-  showTwitterSharePrompt(href, quote);
 }
 
 /**
